libbpf: provide kernel enums with size old kernels
When building against old UAPI headers, libbpf currently only forward
declares enum bpf_stats_type. Unfortunately, C++ compilers require a
size when building against a new enough standard, so we need to provide
one in the forward declaration.
(See https://github.com/libbpf/libbpf/issues/249)

However, because having a size would disagree with the kernel
declaration when present, we have to gate this via LINUX_VERSION_CODE.

Separately but relatedly, bpf_func_linkage needs this treatment as well,
that one is a straight up omission.

I will send these to upstream, though we need to decide if we want to
carry this patch and support older libbpf or if we want to bump the
version outright.